requirement.txt
asgiref==3.5.2
Django==4.1.3
djangorestframework==3.14.0
importlib-metadata==5.0.0
Markdown==3.4.1
python-dotenv==0.21.0
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.6
zipp==3.10.0
